requirements.txt
mysql==0.0.3
mysqlclient==2.1.0
PyMySQL==1.0.2
python-decouple==3.6
